src/test/java/yeongwoo/servlet/domain/member/MemberRepositoryTest.java
package yeongwoo.servlet.domain.member;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;


class MemberRepositoryTest {

    MemberRepository memberRepository = MemberRepository.getInstance();

    @AfterEach
    void afterEach() {
        // 테스트가 끝날 때마다 저장소를 비운다.
        memberRepository.cleanStore();
    }

    @Test
    void save() {
        // given
        Member member = new Member("hello", 20);

        // when
        Member savedMember = memberRepository.save(member);

        // then
        Member findMember = memberRepository.findById(savedMember.getId());
        Assertions.assertThat(findMember).isEqualTo(savedMember);
    }

    @Test
    void findAll() {
        // given
        Member member1 = new Member("member1", 20);
        Member member2 = new Member("member2", 30);

        memberRepository.save(member1);
        memberRepository.save(member2);

        // when
        // then
        Assertions.assertThat(memberRepository.findAll().size()).isEqualTo(2);
        Assertions.assertThat(memberRepository.findAll()).contains(member1, member2);
    }

}
